goma_link_tests: Add test for -gsplit-dwarf
Support for -gsplit-dwarf was added to distributed ThinLTO in
crrev.com/c/2101390. This adds an integration test to verify that
this does indeed cause debug information to be stored in a .dwo
file and not be contained in the main linker output.

Bug: 1058525
Change-Id: I035f4c57ada4efc747a3be4797d7c608da180fcb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2111062
Commit-Queue: Bob Haarman <inglorion@chromium.org>
Reviewed-by: Hans Wennborg <hans@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#752092}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 962e30d835d458cd10e592285cb9404072913b29