Support for auto-generating input for --exclude-fields cmdline param.
After this CL, the rewriter may be run twice, in the following fashion:
1. First run will gather a blocklist of fields to ignore
   1.1. edit/replacement directives from this run will be discarded
2. Second run will consume the blocklist of fields to ignore
   (via --exclude-fields cmdline parameter - see r771819)
   2.1. edit/replacement directives from this run will be applied
        to the source code
   2.2. filters emitted by this run will be discarded

This CL adds code to emit additional output from the rewriter - the list
of fields that should be ignored, because they are used as in-out
arguments.  See "In-out function parameter" section in the design doc
[1] for rationale for ignoring such fields.  See the
tests/in-out-arg-expected.txt file added in this CL for example output.

[1] https://docs.google.com/document/d/1chTvr3fSofQNV_PDPEHRyUgcJCQBgTDOOBriW9gIm9M

Bug: 1069567
Change-Id: I792bb85677ca27f543a3c6da2922cdf8cf991e24
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2161774
Reviewed-by: Bartek Nowierski <bartekn@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Łukasz Anforowicz <lukasza@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#776299}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fbaed2c0c984968b2bdff8be6773d7f40b984589