Remove stale, automatic exclusions from fields-to-ignore.txt
Before this CL, the list of fields to ignore that we've gathered so far
in //tools/clang/rewrite_raw_ptr_fields/fields-to-ignore.txt had two
sections:
1) fields manually added to the list
2) fields automatically added to the list, during one of previous runs
   of the rewriter.

This CL removes the second section of the file, because:
A) that section has been growing stale
B) it is easy to accidentally include the stale section during a rewrite.

When we are actually going to start the SomeClass ->
CheckedPtr<SomeClass> rewrite, *then* it should be okay to once again
save the automatically generated fields-to-ignore.txt (preserving
knowledge of the fields that were actually excluded at the time of the
rewrite).

Bug: 1069567
Change-Id: I9d0a9f411709d7c8a2caebc16e64ef1695a9f410
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2247143
Auto-Submit: Łukasz Anforowicz <lukasza@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Bartek Nowierski <bartekn@chromium.org>
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#779255}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c99e7151e2623b3833a29c87a131cbed08b8296f