Add fallback for Ninja path for clang compile DB
GetNinjaPath() currently hardcodes the path to ninja to be within
//third_party/depot_tools. While this works for Chromium, ANGLE doesn't
include depot_tools in its //third_party. This CL adds a fallback to
essentially perform a 'which ninja' if ninja is not in
//third_party/depot_tools/. This is done with a call to shutil.which().

Additionally, the shutil is specific to Python 3, so both compile_db.py
and generate_compdb.py are updated to '#!/usr/bin/env python3'.

Bug: chromium:1160089
Change-Id: I9cc1266a776b229ee85511254846e01a838e0f01
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2596791
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#838360}
GitOrigin-RevId: caae3361096951c6b26cb6f6d9d75fce4111fc33