Force include blink common/ dirs in the rewrite_raw_ptr_fields rewrite
The code in blink/public/common and blink/common dirs must be included in the rewrite as they are used from the browser process.

This CL adds a syntax where a line starting with ! can be used to negate a file marked for exclusion from the rewrite.

Change-Id: I63b302bc4bf2aa6e88394cd1d0c23ea7750516a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2671441
Commit-Queue: Keishi Hattori <keishi@chromium.org>
Reviewed-by: Łukasz Anforowicz <lukasza@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#850544}
GitOrigin-RevId: 4faf41592dd3626d27512001639d84fe69b6885f