clang build script: omit libc++ tests on mac again after #852778
cycle time increased from 4h15m to 5h after that change. Try to get
most of that back. libc++ tests are slow and we don't use libc++
in production (other than using its headers to build the production
clang binary, but we run the clang test suite already to check that
that works).

Bug: none
Change-Id: I9cd573d953da3f26486b5606b3acff2c608e7885
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2691146
Reviewed-by: Hans Wennborg <hans@chromium.org>
Cr-Commit-Position: refs/heads/master@{#853219}
GitOrigin-RevId: 78292596784d5db38ef9b7f03e6969e5ca50524a