[Include analysis] Use Lengauer-Tarjan for dominance analysis
This makes the "added size" step for a chrome build take 250 s instead
of 2,430 s.

The idea is that this will make the analysis fast enough to handle
larger graphs, such as when analyzing the "added size" of each
individual include edge.

Bug: 1192087
Change-Id: Ic6cf9da271f486e410728d08b51bf54c5775926b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2843628
Commit-Queue: Hans Wennborg <hans@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#875705}
NOKEYCHECK=True
GitOrigin-RevId: 67067a83a3dbacfc848c46ff7af9c7a7054ba6ca