mac, ios: Change install name of asan and ubsan dylibs
This moves us to the default darwin sanitizer setup that was put in
place upstream in https://reviews.llvm.org/D6018 (in Nov 2014!).

Before, the install name of the sanitizer runtime libs was
@executable_path/libclang_rt.asan_osx_dynamic.dylib (etc for ubsan, ios).
That means that when the linker linked an executable against the dylib,
it stored in the executable that it should look for the dylib next to
the executable. The build then copied the dylib into the gn out directory
so that it was next to all executables. (This is also important for
running asan-enabled tests on swarming.)

Now, the (upstream default) install name of the sanitizer runtime libs is
@rpath/libclang_rt.asan_osx_dynamic.dylib (etc). That means that when the
linker links an executable against the dylib, it stores in the executable
to find the dylib on the executable's rpath -- and the clang driver also
tells the linker to put two rpaths in the executable:
1. @executable_path
2. An absolute path to the compiler's resource directory containing the
   dylib

Due to the first entry, the dynamic linker still looks next to the
executable, and in chromium's build it will be there and things should
work just as before (but with slightly less code).

Due to the second entry, building small local one-off binaries by
invoking `third_party/llvm-build/Release+Asserts/bin/clang -fsanitize=address`
will now work too, even if the one-off binary isn't in the
gn out directory.

Bug: 344836
Change-Id: I1f5a86fbdbdb7d22c9629011a395d378245f188b
Cq-Include-Trybots: luci.chromium.try:mac_chromium_asan_rel_ng
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2971579
Commit-Queue: Nico Weber <thakis@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Hans Wennborg <hans@chromium.org>
Cr-Commit-Position: refs/heads/master@{#894234}
NOKEYCHECK=True
GitOrigin-RevId: a04330ab5a23d0e4a60e412a5809ca726a0e4a88