Only generate exclusions for `reinterpret_cast` of *trivial* types.
Since https://crrev.com/c/2648767 the rewriter automatically excludes
fields of SomeStruct if there was a `reinterpret_cast<SomeStruct*>`
somewhere in the code.  This exclusion is

1) very broad - e.g. it excludes fields of any type that has a
   base::WeakPtrFactory as a member (because
   WeakPtrFactory::GetWeakPtr() contains `reinterpret_cast<T*>(...)`).

2) unnecessarily broad - we only want to exclude structs that could
   be safely reinterpret_cast before the rewrite, but aren't safe
   to reinterpret_cast after the rewrite (because the reinterpretation
   skips running CheckedPtr constructors which leads to unbalanced
   BackupRefPtr ref-counts).  In particular, we don't need to exclude
   structs that have non-trivial constructors, because skipping such
   constructors would have been a bug even before the rewrite.

After this CL, the rewriter consults `clang::CXXRecordDecl::isTrivial`
to decide if a struct needs to be excluded from the rewrite.  The CL
also renames the name/tag of the exclusion from
"reinterpret-cast-struct" to "reinterpret-cast-trivial-type".

After this CL, the rewriter generates only 17 reinterpret_cast-related
exclusions for all of Chromium.

Bug: 1069567, 1165613
Change-Id: I2f9d7e527cb7c1295439e7d12cf3664b4c400f14
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3017409
Commit-Queue: Łukasz Anforowicz <lukasza@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#901026}
NOKEYCHECK=True
GitOrigin-RevId: 0c5219fc093415f4692860fb36c9a13a5f21130d