Make the Clang update.py script require Python 3
in preparation for switching the package compression from gzip to xz,
the latter of which relies on Python 3. This is meant to flush out any
remaining invocations of this script under Python 2.

If this breaks a bot, please revert with a link to the broken build
in the revert message.

If this breaks a local use case, please leave a comment on the review
and describe the use case and breakage.

Bug: 1261812
Change-Id: Ie6864bca3e4a045ba454cee4b1ae39db316f63ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3237067
Commit-Queue: Hans Wennborg <hans@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#934973}
NOKEYCHECK=True
GitOrigin-RevId: bebc123d30222bf392d2014e80215e95040b83a4