[clang] Link against libstdc++ statically
This allows us to not have to worry about copying libstdc++.so everywhere we want to run a LLVM binary.

First we'll do a clang roll where we still include libstdc++.so so we can incrementally remove rules/tests that copy libstdc++.so. Then after cleaning up Chrome we can remove libstdc++.so.

Spot checked a couple of Clang binaries and compiler-rt unit tests to make sure they didn't depend on libstdc++.so.

Size of previous vs this clang package:
prev: 202M
this: 206M
prev/bin/clang: 70161760
this/bin/clang: 70640656

libstdc++.so is 1.8M which we can save after removing it from the package later.

Bug: 1286289
Change-Id: I9f10633f5a19eaebcddac1654602a5deec64426b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3513603
Reviewed-by: Hans Wennborg <hans@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Commit-Queue: Arthur Eubanks <aeubanks@google.com>
Cr-Commit-Position: refs/heads/main@{#981214}
NOKEYCHECK=True
GitOrigin-RevId: 10a02a75aac6bce83f7ca7fef8aac8ed9422de10