[Sysroot] Switch to Debian Bullseye stable
The Sid sysroot hasn't been updated in a while, so this switch is
actually an upgrade.

I'm making this change because it's become too difficult to maintain
the Sid sysroot.  The most problematic package is libc6, which seems
to require a new hack on every update.  In the most recent sysroot
upgrade that I tried, glibc began moving pthread symbols from libpthread
to libc.  This causes the binary to fail to load on older systems that
will look for missing pthread symbols in libc.  (Note that old binaries
are forward-compatible on newer systems since glibc added a hack to
libdl to fallback to looking in libc for missing pthread symbols).  To
solve this issue in Chrome, we'd either have to add a hack to lld, or
add a compatibility libpthread that dlopen()s the correct libpthread.

Initially I tried downgrading the sysroot, but that didn't work because
of reliance on newer features (eg. WebRTC depends on pipewire 0.3 which
is only available in Bullseye).

The sysroot was initially upgraded because
1. ChromeOS needed it
    - ChromeOS can use a separate sysroot in the future
2. The Build From Tarball bot needed it for unbundling
    - That bot hasn't been green for months
    - We can leave the necessary libraries bundled
3. We needed a newer libstdc++ with C++14 support
    - We're using libc++ now, so this no longer applies

The current plan is to keep the sysroot at Bullseye until we eventually
drop support for Bullseye.

R=thestig
Bug: None

Change-Id: If1b9c50f2ca343d6c41bcc7b3b69246ca407e2eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3561495
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Commit-Queue: Thomas Anderson <thomasanderson@chromium.org>
Cr-Commit-Position: refs/heads/main@{#988777}
NOKEYCHECK=True
GitOrigin-RevId: 88980b3b5014b877735eda4ea9eaab2c5d561f92