clang: Use gcc package built for bionic as host gcc
Now that linux_upload_clang is on bionic, we can build gcc it
downloads on bionic too.

I built this gcc in a ubuntu:bionic image as described here:
https://bugs.chromium.org/p/chromium/issues/detail?id=1302636#c18

The clang build uses an older clang as host compiler for building
the bootstrap compiler (on the packaging bots) or the final
compiler (on the ToT bots), but we still use the libstdc++ from
the gcc image at the moment. (It's statically linked into
the final clang, and some rust tooling uses that libstdc++ too).

The clang we build already requires bionic (since linux_upload_clang
runs bionic, and clang isn't' built in a sysroot yet), so
libstdc++ requiring bionic shouldn't cause any new issues.

The current (trusty-based) gcc had header issues that made
it impossible to use it with a sysroot for building clang.
This is better with gcc targeting a slightly newer linux
(or, rather, glibc). So this change hopefully helps with
eventually building clang itself against a sysroot.

Hopefully no behavior change.

Bug: 1302636
Change-Id: Ib65a7135418985c0a53d23e62682d1bd5e26ca86
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3650696
Reviewed-by: Hans Wennborg <hans@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1004473}
NOKEYCHECK=True
GitOrigin-RevId: 1b307efd06d29068c5fac51b539a98d4989b9f99