Stop compile_db_test.py from running on Python 2
compile_db_test.py has been running on Python 3 for a long time, but it
is also running on Python 2, which is unnecessary. This change turns off
Python 2 for this script. The change was tested with this command:

  git cl presubmit --files tools/clang/pylib/clang/compile_db_test.py

Bug: 1313804
Change-Id: Ib35801ed5e1ad1113df762b1da9276c14c78b0f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3687784
Auto-Submit: Bruce Dawson <brucedawson@chromium.org>
Reviewed-by: Amy Huang <akhuang@google.com>
Commit-Queue: Amy Huang <akhuang@google.com>
Cr-Commit-Position: refs/heads/main@{#1010716}
NOKEYCHECK=True
GitOrigin-RevId: fd48865d76b64a6e8b20046d8b387a7231bae951