[clang] Use runtimes build to build compiler-rt/libcxx
libcxx has been moving off of LLVM_ENABLE_PROJECTS for a while, and compiler-rt seems to also be headed in that direction [1].

This also simplifies our build script since the runtimes CMake logic handles much of the same thing.

-DLLVM_RUNTIME_TARGETS contains a list of target triples that will participate in the runtimes logic. CMake arguments for these sub-CMake invocations are passed via -DRUNTIMES_$target_$cmakevar.

compiler-rt often depends on the builtins library (which itself is in compiler-rt). The runtimes logic handles this with -DLLVM_BUILTIN_TARGETS/-DBUILTINS_$target_$cmakevar. The builtins are built first before the rest of compiler-rt.

If -DLLVM_RUNTIME_TARGETS is omitted (e.g. in the macOS case), it's as if specifying the target platform.

If -DLLVM_RUNTIME_TARGETS is specified, -DLLVM_ENABLE_RUNTIMES is still required to be not empty and is only used to calculate what dependencies (e.g. opt) should be built before running the sub-invocations.

For macOS, see [2] on why we don't need to specify multiple targets for ios/etc.

[1] https://discourse.llvm.org/t/rfc-deprecate-and-remove-llvm-build-external-compiler-rt/62058/9
[2] https://github.com/llvm/llvm-project/blob/52818fd97f0f2b13d12e66de9f06b9f4cbc0be07/llvm/runtimes/CMakeLists.txt#L47

Change-Id: I9eccd2827e54743a94e425e5d1cfc94358fa926e
Bug: 1258656
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3690694
Reviewed-by: Hans Wennborg <hans@chromium.org>
Commit-Queue: Arthur Eubanks <aeubanks@google.com>
Cr-Commit-Position: refs/heads/main@{#1013287}
NOKEYCHECK=True
GitOrigin-RevId: a2a9cf6cfa3aec3410e2053117990f6954fb3a4b