[clang] Fix the pinned clang version used for bootstrapping
Using the version from HEAD~ was elegant but didn't work when run
outside a Git repository or when HEAD~ contained changes to the clang
version.

Just using a fixed version which we can bump every once in a while is
less elegant, but it avoids these problems and is also a little simpler.

Bug: 1346047
Change-Id: I5f79b4b39829fa074f321c7d5b1c3b103837a8b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3791908
Reviewed-by: Arthur Eubanks <aeubanks@google.com>
Commit-Queue: Hans Wennborg <hans@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1031544}
NOKEYCHECK=True
GitOrigin-RevId: 0a6813ad98384f23e9874c142aa08141cc22f216