Issue warning for default malloc used in blink
The default malloc is discouraged in blink code. This CL adds a
checker in blink_gc_plugin to check for classes that use the default
malloc.

The checker is not enabled by default yet. We'll enable it for blink
subdirectories incrementally and fix the issues found.

Bug: 1357410
Change-Id: I50906a89d767b955a72a1f6bfce4ad5a2ef4159a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3860855
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Auto-Submit: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1040240}
NOKEYCHECK=True
GitOrigin-RevId: fa3b24999ef2ba988906737a5433626c0abdb6d7