RecordInfo::IsConsideredAbstract() with implicit default constructor
This CL let RecordInfo::IsConsideredAbstract() return false if the
class has the implicit default constructor. In the case the class
can be instantiated so should be considered in the manual dispatch
and finalize methods.

This change is not behind a flag because the plugin didn't find any
new problem is the current blink code.

Change-Id: I80a6c87b2cd9f83e710923c795866b3779a0ed5b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3861170
Commit-Queue: Kentaro Hara <haraken@chromium.org>
Auto-Submit: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1040260}
NOKEYCHECK=True
GitOrigin-RevId: bb2f325277339b5cc378d2fd9f7c3d92b4446655